pyproject.toml
[tool.poetry]
name = "django-cars-project"
version = "0.1.0"
description = ""
authors = ["Anastasiia Mirovska <anastasiia.mirovska@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.4"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
drf-yasg = "^1.21.7"
setuptools = "^74.1.2"
celery = "^5.4.0"
redis = "^5.0.8"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"



[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
